requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography>=41.0.5,<43
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Dance==7.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyOpenSSL==24.0.0
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.0.1
WTForms==3.2.1
